+++ .gitignore
@@ -2,4 +2,5 @@
 *venuspy.egg-info
 /dist
 .vscode
+/venv
